+++ src/app/loom/page.tsx
@@ -1,7 +1,7 @@
 
 'use client';
 
-import { useState, useEffect, useCallback, useRef } from 'react';
+import { useState, useEffect, useCallback } from 'react';
 import { CanvasZone } from '@/app/loom/components/canvas/canvas-zone';
 import { PalettePanel } from '@/app/loom/components/panels/palette-panel';
 import { InspectorPanel } from '@/app/loom/components/panels/inspector-panel';
@@ -15,7 +15,6 @@ import { useToast } from '@/hooks/use-toast';
 import { ResizableHorizontalPanes } from '@/app/loom/components/layout/resizable-horizontal-panes';
 import { ResizableVerticalPanes } from '@/app/loom/components/layout/resizable-vertical-panes';
 import { TooltipProvider } from '@/components/ui/tooltip';
-import { useBeepChat } from '@/hooks/use-beep-chat';
 import eventBus from '@/lib/event-bus';
 import { exampleTemplates } from '@/app/loom/data/templates';
 import { useLoomStore } from '@/stores/loom.store';
@@ -23,16 +22,16 @@ import { cn } from '@/lib/utils';
 import { shallow } from 'zustand/shallow';
 import { BottomBar } from './components/layout/bottom-bar';
 import { useActionRequestStore } from '@/stores/action-request.store';
+import { useBeepChatStore } from '@/stores/beep-chat.store';
+import { useLoomOrchestrator } from '@/hooks/use-loom-orchestrator';
 
 
 import type {
   WorkflowNodeData,
-  NodeStatus,
   PanelVisibility,
   ConnectingState,
   AiGeneratedFlowData,
   ConsoleMessage,
-  TimelineEvent,
   WorkflowTemplate,
   Connection,
 } from '@/types/loom';
@@ -44,13 +43,14 @@ export default function LoomStudioPage() {
     workflowName, nodeExecutionStatus,
     setWorkflow, setSelectedNodeId, updateNode, deleteNode, addNode, addConnection,
     clearWorkflow, addConsoleMessage, addTimelineEvent, clearConsole,
-    setNodeExecutionStatus, updateNodeStatus,
   } = useLoomStore(state => ({...state}), shallow);
   
   const { requests: actionRequests, resolveActionRequest } = useActionRequestStore(state => ({
     requests: state.requests,
     resolveActionRequest: state.resolveActionRequest,
   }), shallow);
+
+  const { isWorkflowRunning, handleRunWorkflow, handleRunNode } = useLoomOrchestrator();
 
   const [connectingState, setConnectingState] = useState<ConnectingState | null>(null);
   const [panelVisibility, setPanelVisibility] = useState<PanelVisibility>({
@@ -68,14 +68,11 @@ export default function LoomStudioPage() {
     error: true,
   });
   const [isTemplateSelectorOpen, setIsTemplateSelectorOpen] = useState(false);
-  const [isWorkflowRunning, setIsWorkflowRunning] = useState(false);
-
-  const { append: beepAppend, messages: beepMessages, isLoading: isBeepLoading, error: beepError } = useBeepChat();
+  
   const isMobile = useIsMobile();
   const { toast } = useToast();
   
   const selectedNode = nodes.find(n => n.id === selectedNodeId) || null;
-  const prevNodeStatusRef = useRef<Record<string, NodeStatus>>({});
 
   useEffect(() => {
     addConsoleMessage('info', 'Loom Studio Initialized. AI is standing by to generate workflows.');
@@ -119,226 +116,6 @@ export default function LoomStudioPage() {
     };
   }, [handleFlowGenerated]);
   
-  const runningNodeId = Object.keys(nodeExecutionStatus).find(
-    (id) => nodeExecutionStatus[id] === 'running'
-  );
-
-  useEffect(() => {
-    if (beepError && runningNodeId) {
-      const runningNode = nodes.find(n => n.id === runningNodeId);
-      if (runningNode) {
-        const errorMessage = beepError.message || 'An unknown API error occurred.';
-        updateNode(runningNode.id, {
-          status: 'failed',
-          config: { ...runningNode.config, output: { error: errorMessage } }
-        });
-        updateNodeStatus(runningNode.id, 'failed');
-        addTimelineEvent({
-          type: 'node_failed',
-          message: `Node failed due to API error: ${errorMessage}`,
-          nodeId: runningNode.id,
-          nodeTitle: runningNode.title,
-        });
-        addConsoleMessage('error', `Node ${runningNode.title} failed: ${errorMessage}`);
-      }
-    }
-  }, [beepError, runningNodeId, nodes, updateNode, updateNodeStatus, addTimelineEvent, addConsoleMessage]);
-
-  const handleRunNode = useCallback(async (nodeId: string, inputData?: any) => {
-    const nodeToRun = nodes.find(n => n.id === nodeId);
-    if (!nodeToRun) return;
-
-    if (nodeToRun.config?.beepEmotion) eventBus.emit('beep:setEmotion', nodeToRun.config.beepEmotion);
-
-    updateNodeStatus(nodeId, 'running');
-    addTimelineEvent({ type: 'node_running', message: `Executing node: ${nodeToRun.title}`, nodeId, nodeTitle: nodeToRun.title });
-    
-    let prompt = '';
-    const promptInput = inputData ? `\n\nUse the following data as input for your task:\n${JSON.stringify(inputData)}` : '';
-    
-    switch (nodeToRun.type) {
-      case 'web-summarizer':
-        prompt = `Please summarize the content from this URL: ${nodeToRun.config?.url}`;
-        break;
-      case 'prompt':
-      case 'agent-call':
-        prompt = (nodeToRun.config?.promptText || `Execute generic prompt for node ${nodeToRun.title}`) + promptInput;
-        break;
-      case 'data-transform':
-        prompt = `Please execute a data transformation with the following logic: "${nodeToRun.config?.transformationLogic}".` + promptInput;
-        break;
-      case 'conditional':
-        prompt = `Please evaluate the condition: "${nodeToRun.config?.condition}". Return only a JSON object with a single boolean 'result' key, like \`{"result": true}\`.` + promptInput;
-        break;
-      default:
-        toast({ title: "Execution Not Implemented", description: `Backend for '${nodeToRun.type}' is not yet implemented.`});
-        addConsoleMessage('warn', `Execution for node type '${nodeToRun.type}' is not implemented. Faking failure.`);
-        setTimeout(() => updateNodeStatus(nodeId, 'failed'), 1500);
-        return;
-    }
-    if (prompt) {
-      addConsoleMessage('info', `Dispatching task to BEEP for node "${nodeToRun.title}": ${prompt}`);
-      beepAppend({ role: 'user', content: prompt });
-    }
-  }, [nodes, addConsoleMessage, addTimelineEvent, toast, beepAppend, updateNodeStatus]);
-
-  // Centralized hook for processing agent results and updating node state
-  useEffect(() => {
-    if (isBeepLoading || !runningNodeId) {
-        return;
-    }
-    
-    const lastMessage = beepMessages[beepMessages.length - 1];
-    if (!lastMessage) return;
-
-    const runningNode = nodes.find(n => n.id === runningNodeId);
-    if (!runningNode) return;
-
-    let resultOutput: any = null;
-    let nodeSucceeded = false;
-
-    // Case 1: The last message is a tool result for a server-side tool.
-    if (lastMessage.role === 'tool') {
-        try {
-            resultOutput = JSON.parse(lastMessage.content as string);
-            if(resultOutput.isClientSide) {
-                // This was a client-side tool call, not the end of a server node execution.
-                resultOutput = null; 
-            } else {
-               nodeSucceeded = resultOutput?.error ? false : true;
-            }
-        } catch (e) {
-            resultOutput = { error: "Failed to parse tool result." };
-            nodeSucceeded = false;
-        }
-    }
-    // Case 2: The last message is a simple text response from the assistant, meaning the node is done.
-    else if (lastMessage.role === 'assistant' && !lastMessage.tool_calls?.length) {
-        let content = lastMessage.content;
-        try {
-          // Attempt to parse content as JSON for conditional nodes
-          const parsedJson = JSON.parse(content);
-          resultOutput = parsedJson;
-        } catch (e) {
-          // Fallback for non-JSON text responses
-          resultOutput = { result: content };
-        }
-        nodeSucceeded = true;
-    }
-    
-    // If we have a conclusive result for the running node, update its state.
-    if (resultOutput) {
-        const newStatus = nodeSucceeded ? 'completed' : 'failed';
-        updateNode(runningNode.id, {
-            status: newStatus,
-            config: { ...runningNode.config, output: resultOutput }
-        });
-        updateNodeStatus(runningNode.id, newStatus);
-        addTimelineEvent({
-            type: nodeSucceeded ? 'node_completed' : 'node_failed',
-            message: nodeSucceeded ? `Node finished successfully.` : `Node failed: ${resultOutput.error || 'Unknown error'}`,
-            nodeId: runningNode.id,
-            nodeTitle: runningNode.title,
-        });
-    }
-  }, [beepMessages, isBeepLoading, runningNodeId, nodeExecutionStatus, nodes, updateNode, updateNodeStatus, addTimelineEvent, addConsoleMessage]);
-
-  useEffect(() => {
-    if (!isWorkflowRunning) {
-      prevNodeStatusRef.current = {};
-      return;
-    }
-  
-    const newlyCompletedNodes = nodes.filter(node => 
-      nodeExecutionStatus[node.id] === 'completed' &&
-      prevNodeStatusRef.current[node.id] !== 'completed'
-    );
-  
-    if (newlyCompletedNodes.length > 0) {
-      newlyCompletedNodes.forEach(completedNode => {
-        const allOutgoingConnections = connections.filter(c => c.from === completedNode.id);
-        if (allOutgoingConnections.length === 0) return;
-  
-        let connectionsToFollow: Connection[] = [];
-  
-        if (completedNode.type === 'conditional') {
-          const result = !!completedNode.config?.output?.result; // Safely coerce to boolean
-          addConsoleMessage('info', `Conditional node "${completedNode.title}" evaluated to: ${result}`);
-  
-          if (result && allOutgoingConnections.length > 0) {
-            // True path is the first connection by convention
-            connectionsToFollow.push(allOutgoingConnections[0]);
-          } else if (!result && allOutgoingConnections.length > 1) {
-            // False path is the second connection by convention
-            connectionsToFollow.push(allOutgoingConnections[1]);
-          }
-        } else {
-          // For all other nodes, trigger all outgoing paths
-          connectionsToFollow = allOutgoingConnections;
-        }
-  
-        connectionsToFollow.forEach(conn => {
-          const nextNode = nodes.find(n => n.id === conn.to);
-          if (nextNode) {
-            updateNodeStatus(nextNode.id, 'queued');
-            handleRunNode(nextNode.id, completedNode.config?.output);
-          }
-        });
-      });
-    }
-    
-    prevNodeStatusRef.current = { ...nodeExecutionStatus };
-  
-    const isFinished = !Object.values(nodeExecutionStatus).some(
-      status => status === 'running' || status === 'queued'
-    );
-    
-    const hasStarted = Object.keys(nodeExecutionStatus).length > 0;
-  
-    if (hasStarted && isFinished) {
-      setIsWorkflowRunning(false);
-      const hasFailures = Object.values(nodeExecutionStatus).some(s => s === 'failed');
-      toast({ 
-        title: 'Workflow Finished', 
-        description: `Execution of "${workflowName}" is complete.`,
-        variant: hasFailures ? 'destructive' : 'default',
-      });
-      addTimelineEvent({ type: hasFailures ? 'workflow_failed' : 'workflow_completed', message: 'Workflow execution finished.' });
-    }
-  
-  }, [nodeExecutionStatus, isWorkflowRunning, nodes, connections, updateNodeStatus, handleRunNode, addTimelineEvent, toast, workflowName, addConsoleMessage]);
-
-
-  const handleRunWorkflow = useCallback(() => {
-    if (nodes.length === 0) {
-      toast({ title: "Empty Workflow", description: "Add nodes to the canvas before running.", variant: 'destructive' });
-      return;
-    }
-    setIsWorkflowRunning(true);
-    addConsoleMessage('info', `User initiated workflow execution for "${workflowName}".`);
-    addTimelineEvent({ type: 'workflow_start', message: `Workflow "${workflowName}" started.` });
-    
-    const initialStatuses: Record<string, NodeStatus> = {};
-    nodes.forEach(n => initialStatuses[n.id] = 'pending');
-    
-    const startNodes = nodes.filter(n => !connections.some(c => c.to === n.id));
-    
-    if (startNodes.length === 0 && nodes.length > 0) {
-        toast({ title: "Execution Error", description: "No starting node found. Check for circular dependencies.", variant: 'destructive' });
-        setIsWorkflowRunning(false);
-        return;
-    }
-
-    startNodes.forEach(n => {
-        initialStatuses[n.id] = 'queued';
-        addTimelineEvent({ type: 'node_queued', message: `Starting node "${n.title}" queued.`, nodeId: n.id, nodeTitle: n.title });
-        handleRunNode(n.id);
-    });
-
-    setNodeExecutionStatus(initialStatuses);
-
-  }, [nodes, connections, workflowName, addConsoleMessage, addTimelineEvent, toast, handleRunNode, setNodeExecutionStatus]);
-
   const handleNodeDropped = (newNodeData: Omit<WorkflowNodeData, 'id' | 'status'> & { status?: NodeStatus }) => {
     const newNode = addNode(newNodeData);
     setSelectedNodeId(newNode.id);
@@ -429,6 +206,7 @@ export default function LoomStudioPage() {
   const handleAgentActionResponse = (requestId: string, responseStatus: 'approved' | 'denied' | 'responded', details?: string) => {
     addConsoleMessage('info', `User responded to action request ${requestId} with status: ${responseStatus}`);
     
+    const { append: beepAppend } = useBeepChatStore.getState();
     resolveActionRequest(requestId, responseStatus, details);
 
     let userResponseText = `User response for request ${requestId}: ${responseStatus}.`;
@@ -508,5 +286,3 @@ export default function LoomStudioPage() {
     </div>
   );
 }
-
-
